requirements.txt
tensorflow==2.12.0
numpy==1.24.4
opencv-python-headless==4.7.0.72
Pillow==9.5.0
requests==2.31.0
h5py==3.8.0
